+++ src/icloudpd/server/static/js/toast.js
@@ -1,0 +1,6 @@
+document.addEventListener("htmx:afterRequest", function(evt) {
+    if (evt.detail.xhr.status >= 400 && evt.detail.xhr.status < 600) {
+        const toastBoostrap = bootstrap.Toast.getOrCreateInstance(document.getElementById('toasts'));
+        toastBoostrap.show();
+    }
+});